Рефакторинг ч. 1
